+++ spec/features/users_can_create_involuntary_conversion_projects_spec.rb
@@ -39,6 +39,9 @@ RSpec.feature "Users can create new involuntary conversion projects" do
 
       fill_in "Handover comments", with: "A new handover comment"
 
+      within("#directive-academy-order") do
+        choose "Yes"
+      end
       click_button("Continue")
 
       expect(page).to have_content(I18n.t("project.show.title"))